fix: 헤더에 게시판 카테고리를 헤더 컴포넌트에서 분리
