+++ pages/[boardId]/index.tsx
@@ -2,6 +2,7 @@ import React, { useState } from 'react';
 import styled from 'styled-components';
 
 import Header from '@/src/components/Header';
+import BoardCategories from '@/src/components/BoardCategories';
 import CommonRightLayout from '@/src/views/common/commonRightLayout';
 
 import BoardView from '@/src/views/boardView';
@@ -12,7 +13,9 @@ const FooterSection = styled.div`
   bottom: 0;
 `;
 
-const Container = styled.div``;
+const Container = styled.div`
+  ${({ theme }) => theme.containers.mainContainer}
+`;
 
 const LayoutContainer = styled.div`
   width: 1125px;
@@ -96,6 +99,7 @@ function BoardPage() {
   return (
     <Container>
       <Header />
+      <BoardCategories />
       <LayoutContainer>
         <BoardView
           contentList={contentList}
