+++ pages/my-article/index.tsx
@@ -4,6 +4,7 @@ import styled from 'styled-components';
 /** components */
 import MyArticleView from '@/src/views/myArticleView';
 import Header from '@/src/components/Header';
+import BoardCategories from '@/src/components/BoardCategories';
 import Footer from '@/src/components/Footer';
 import CommonRightLayout from '@/src/views/common/commonRightLayout';
 
@@ -12,7 +13,9 @@ const FooterSection = styled.div`
   bottom: 0;
 `;
 
-const Container = styled.div``;
+const Container = styled.div`
+  ${({ theme }) => theme.containers.mainContainer}
+`;
 
 const LayoutContainer = styled.div`
   width: 1125px;
@@ -25,6 +28,7 @@ function myArticle() {
   return (
     <Container>
       <Header />
+      <BoardCategories />
       <LayoutContainer>
         <MyArticleView />
         <CommonRightLayout />
